chore: move balancer-v3 abis to `src/abi/balancer-v3` folder
